+++ packages/manager/modules/pci/src/projects/project/storages/databases/database/service-integration/service-integration.component.js
@@ -1,0 +1,20 @@
+import controller from './service-integration.controller';
+import template from './service-integration.html';
+
+export default {
+  bindings: {
+    projectId: '<',
+    database: '<',
+    trackDashboard: '<',
+    goToAddServiceIntegration: '<',
+    goToDeleteServiceIntegration: '<',
+    refreshServiceIntegration: '<',
+    serviceIntegrationList: '<',
+    kafkaServicesList: '<',
+    addableServicesList: '<',
+    stopPollingIntegrationsStatus: '<',
+    replicationsList: '<',
+  },
+  controller,
+  template,
+};